refactor: split auth.ts controller (1884 lines) into 7 focused modules
Split the monolithic auth controller into focused sub-modules by concern:
- auth-common.ts: shared types, JWT helpers, singletons, validation
- auth-google.ts: Google OAuth web + mobile handlers
- auth-password.ts: password login, registration, email verification, reset
- auth-webauthn.ts: WebAuthn/passkey register, auth, list, delete
- auth-diia.ts: Diia government ID auth + digital signing flows
- auth-oidc.ts: OIDC/Authentik SSO handlers
- auth-profile.ts: user profile, contracts, avatar, banner, logout, refresh

auth.ts is now a barrel re-export so existing importers are unaffected.
